Add files via upload
key.py - simply captures what you are typing and stores it in a text file 

Key_logger_TimeUpDown - Captures timing data the dwell time and fly time the dwell time is how long the key is held down, the fly time is the time between the previous key release and next key press. This code is fun to play with but the results are inaccurate right now.